// src/common/enums/jwt.enum.ts
export enum Jwt {
    ACCESS_TOKEN_SECRET = 'JWT_ACCESS_SECRET',
    ACCESS_TOKEN_EXPIRES = 'JWT_ACCESS_EXPIRES',
    REFRESH_TOKEN_SECRET = 'JWT_REFRESH_SECRET',
    REFRESH_TOKEN_EXPIRES = 'JWT_REFRESH_EXPIRES',
    VERIFY_TOKEN_SECRET = 'JWT_VERIFY_SECRET',
    VERIFY_TOKEN_EXPIRES = 'JWT_VERIFY_EXPIRES',
}